+++ packages/interface/src/components/action-item/index.js
@@ -1,13 +1,12 @@
 /**
  * External dependencies
  */
-import { isArray, isEmpty, noop } from 'lodash';
+import { isEmpty, noop } from 'lodash';
 
 /**
  * WordPress dependencies
  */
 import { ButtonGroup, Button, Slot, Fill } from '@wordpress/components';
-import deprecated from '@wordpress/deprecated';
 import { Children } from '@wordpress/element';
 
 function ActionItemSlot( {
@@ -17,18 +16,6 @@ function ActionItemSlot( {
 	bubblesVirtually,
 	...props
 } ) {
-	if ( isArray( Component ) ) {
-		deprecated(
-			'Passing a tuple of components with `as` prop to `ActionItem.Slot` component',
-			{
-				since: '10.2',
-				plugin: 'Gutenberg',
-				alternative: 'a component with `as` prop',
-				version: '10.3',
-			}
-		);
-		Component = Component[ 0 ];
-	}
 	return (
 		<Slot
 			name={ name }